[Refactor] bottomsheet refactoring
- component 패키지로 이동
- 파라미터 추가
