+++ app/src/main/java/com/and04/naturealbum/ui/maps/MapScreen.kt
@@ -57,6 +57,7 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import coil3.compose.AsyncImage
 import com.and04.naturealbum.R
 import com.and04.naturealbum.data.room.PhotoDetail
+import com.and04.naturealbum.ui.component.CustomBottomSheetComponent
 import com.naver.maps.geometry.LatLng
 import com.naver.maps.map.MapView
 import com.naver.maps.map.clustering.ClusterMarkerInfo
@@ -198,204 +199,6 @@ fun PhotoGrid(
     LazyVerticalGrid(columns = GridCells.Adaptive(256.dp)) {
         items(photos) { photo ->
             AsyncImage(model = photo.photoUri, contentDescription = null)
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun BottomSheetScreen(
-    onBackButtonClick: () -> Unit
-) {
-    val customBottomSheetVisible = remember { mutableStateOf(true) }
-    Box(
-        modifier = Modifier
-            .fillMaxSize(),
-        contentAlignment = Alignment.Center
-    ) {
-//        MapScreen()
-//        Button(onClick = { customBottomSheetVisible.value = true }) {
-//            Box(
-//                modifier = Modifier.size(
-//                    50.dp
-//                )
-//            )
-//        }
-
-        Box(
-            modifier = Modifier
-                .fillMaxSize(),
-        ) {
-            CustomBottomSheetComponent {
-
-            }
-        }
-    }
-}
-
-enum class BottomSheetState {
-    Collapsed,
-    HalfExpanded,
-    Expanded
-}
-
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-fun CustomBottomSheetComponent(
-    initialValue: BottomSheetState = BottomSheetState.Collapsed,
-    modifier: Modifier = Modifier,
-    content: @Composable () -> Unit = {}
-) {
-    val density = LocalDensity.current
-    val decayAnimationSpec = rememberSplineBasedDecay<Float>()
-    val screenHeight = with(density) { LocalConfiguration.current.screenHeightDp.dp.toPx() }
-    val handleHeight = with(density) { 40.dp.toPx() }
-    var currentState by remember { mutableStateOf(initialValue) }
-    val state = remember {
-        AnchoredDraggableState(
-            initialValue = currentState,
-            anchors = DraggableAnchors {
-                BottomSheetState.Collapsed at (screenHeight - handleHeight)
-                BottomSheetState.HalfExpanded at screenHeight * 0.5f
-                BottomSheetState.Expanded at screenHeight * 0.05f
-            },
-            positionalThreshold = { distance: Float -> distance * 0.5f },
-            velocityThreshold = { with(density) { 100.dp.toPx() } },
-            snapAnimationSpec = tween(),
-            decayAnimationSpec = decayAnimationSpec
-        )
-    }
-
-    LaunchedEffect(currentState) {
-        state.animateTo(currentState)
-    }
-    LaunchedEffect(state.currentValue) {
-        currentState = state.currentValue
-    }
-
-    Box(
-        modifier = modifier
-            .offset {IntOffset(0, state.requireOffset().roundToInt())}
-            .fillMaxWidth()
-            .padding(horizontal = 16.dp),
-        contentAlignment = Alignment.Center
-    ) {
-        ElevatedCard(
-            shape = RoundedCornerShape(topStart = 40.dp, topEnd = 40.dp),
-        ) {
-            Icon(
-                imageVector = Icons.Default.DragHandle,
-                contentDescription = null,
-                modifier = modifier
-                    .fillMaxWidth()
-                    .height(36.dp)
-                    .anchoredDraggable(
-                        state = state,
-                        orientation = Orientation.Vertical,
-                    )
-                    .clickable(
-                        interactionSource = remember { MutableInteractionSource() },
-                        indication = null
-                    ) {
-                        currentState = when (currentState) {
-                            BottomSheetState.Collapsed -> BottomSheetState.HalfExpanded
-                            BottomSheetState.HalfExpanded -> BottomSheetState.Collapsed
-                            BottomSheetState.Expanded -> BottomSheetState.HalfExpanded
-                        }
-                    }
-            )
-            Box(
-                modifier = modifier
-                    .fillMaxSize()
-                    .padding(start = 16.dp, end = 16.dp, top = 16.dp),
-                contentAlignment = Alignment.TopCenter
-            ) {
-                content()
-            }
-        }
-    }
-}
-
-
-@Preview
-@Composable
-fun BottomSheetScreenCollapsedPreView(
-) {
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(Color.White),
-        contentAlignment = Alignment.BottomCenter
-    ) {
-        CustomBottomSheetComponent(initialValue = BottomSheetState.Collapsed) {
-            Column {
-                Box(
-                    modifier = Modifier
-                        .size(50.dp)
-                        .background(Color.Blue)
-                )
-                Box(
-                    modifier = Modifier
-                        .size(50.dp)
-                        .background(Color.Black)
-                )
-            }
-        }
-    }
-}
-
-@Preview
-@Composable
-fun BottomSheetScreenHalfExpandedPreView(
-) {
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(Color.White),
-        contentAlignment = Alignment.BottomCenter
-    ) {
-        // BottomSheet
-        CustomBottomSheetComponent(initialValue = BottomSheetState.HalfExpanded) {
-            Column {
-                Box(
-                    modifier = Modifier
-                        .size(50.dp)
-                        .background(Color.Blue)
-                )
-                Box(
-                    modifier = Modifier
-                        .size(50.dp)
-                        .background(Color.Black)
-                )
-            }
-        }
-    }
-}
-
-@Preview
-@Composable
-fun BottomSheetScreenExpandedPreView(
-) {
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(Color.White),
-        contentAlignment = Alignment.BottomCenter
-    ) {
-        // BottomSheet
-        CustomBottomSheetComponent(initialValue = BottomSheetState.Expanded) {
-            Column {
-                Box(
-                    modifier = Modifier
-                        .size(50.dp)
-                        .background(Color.Blue)
-                )
-                Box(
-                    modifier = Modifier
-                        .size(50.dp)
-                        .background(Color.Black)
-                )
-            }
         }
     }
 }
